pix2code/requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==1.5.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-bootstrap4==0.0.8
django-upload-validator==1.0.2
django-widget-tweaks==1.4.5
django-widgets-improved==1.5.0
entrypoints==0.3
enum34==1.1.6
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.7.1
html5lib==0.9999999
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.3
notebook==5.7.8
numpy==1.13.3
opencv-python==3.3.0.10
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.8.0
pydotplus==2.0.2
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.4.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
